fix: address PR review important issues
- Handle clipboard writeText rejection (App.tsx, committed above)
- Log warning for unparseable messages in kill-client
- Log ENOENT once when git not found in session-name
- Extract formatRelativeTime to shared lib/format-time.ts
