+++ packages/web/src/components/chat/SessionSwitcher.tsx
@@ -5,6 +5,7 @@
  * unread badges, and question-pending markers. Tap a session to switch to it.
  */
 
+import { formatRelativeTime } from '@/lib/format-time';
 import type { UISession } from '@/types';
 import type { UUID } from '@remi/shared/types.ts';
 import { clsx } from 'clsx';
@@ -17,26 +18,6 @@ interface SessionSwitcherProps {
   readonly isOpen: boolean;
   readonly onSelectSession: (id: UUID) => void;
   readonly onClose: () => void;
-}
-
-/** Format relative time for session list */
-function formatRelativeTime(timestamp: string): string {
-  const date = new Date(timestamp);
-  const now = new Date();
-  const diffMs = now.getTime() - date.getTime();
-  const diffMins = Math.floor(diffMs / 60000);
-  const diffHours = Math.floor(diffMs / 3600000);
-  const diffDays = Math.floor(diffMs / 86400000);
-
-  if (diffMins < 1) return 'Just now';
-  if (diffMins < 60) return `${diffMins}m ago`;
-  if (diffHours < 24) return `${diffHours}h ago`;
-  if (diffDays < 7) return `${diffDays}d ago`;
-
-  return date.toLocaleDateString(undefined, {
-    month: 'short',
-    day: 'numeric',
-  });
 }
 
 /** Status dot color based on connection and agent status */
