+++ packages/web/src/components/session/SessionCard.tsx
@@ -5,6 +5,7 @@
  */
 
 // Status icons are rendered via StatusDot, not used directly
+import { formatRelativeTime } from '@/lib/format-time';
 import type { AgentStatus, ConnectionStatus, UISession } from '@/types';
 import { clsx } from 'clsx';
 
@@ -12,26 +13,6 @@ interface SessionCardProps {
   readonly session: UISession;
   readonly isActive: boolean;
   readonly onClick: () => void;
-}
-
-/** Format relative time */
-function formatRelativeTime(timestamp: string): string {
-  const date = new Date(timestamp);
-  const now = new Date();
-  const diffMs = now.getTime() - date.getTime();
-  const diffMins = Math.floor(diffMs / 60000);
-  const diffHours = Math.floor(diffMs / 3600000);
-  const diffDays = Math.floor(diffMs / 86400000);
-
-  if (diffMins < 1) return 'Just now';
-  if (diffMins < 60) return `${diffMins}m ago`;
-  if (diffHours < 24) return `${diffHours}h ago`;
-  if (diffDays < 7) return `${diffDays}d ago`;
-
-  return date.toLocaleDateString(undefined, {
-    month: 'short',
-    day: 'numeric',
-  });
 }
 
 /** Status dot with color and animation */
